Refactor: check media exist before delete tags and posts

+++ app/Services/PostService.php
@@ -57,7 +57,9 @@ final class PostService
     {
         // delete related media
         $media = $post->media;
-        $this->mediaService->destroy($media);
+        if ($media) {
+            $this->mediaService->destroy($media);
+        }
 
         // delete its sections
         $sections = $post->sections;
